requirements.txt
tensorflow>=2.4.0
numpy>=1.19.5
pandas>=1.2.0
matplotlib>=3.3.4
seaborn>=0.11.1
arch>=4.19
jquantsapi>=0.1.0
pyyaml>=5.4.1
tqdm>=4.61.0
scikit-learn>=0.24.2
jupyterlab>=3.0.0
python-dotenv>=0.19.0